third_party/blink/web_tests/http/tests/websocket/port-blocking-expected.txt

WebSocket: Test port blocking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
